[#42] Bcrypt에 대한 테스트 목적에 맞게 분리
