+++ src/test/java/com/app/todolist/api/members/MemberServiceTest.java
@@ -65,21 +65,6 @@ class MemberServiceTest {
     }
 
     @Test
-    @DisplayName("회원 생성 시 비밀번호가 암호화되어 생성된다.")
-    void testPasswordEncryption() {
-        // given
-        String password = "1234";
-
-        // when
-        String hashedPassword = BCrypt.hashpw(password, BCrypt.gensalt());
-
-        // then
-        assertThat(hashedPassword).isNotNull();
-        assertThat(password).isNotEqualTo(hashedPassword);
-        assertThat(BCrypt.checkpw(password, hashedPassword)).isTrue();
-    }
-
-    @Test
     @DisplayName("회원 생성 시 비밀번호가 암호화되어 저장된다.")
     void testCreateMember_passwordShouldBeEncrypted() {
         // given
